Change to use CURLFile when tweet with media.
In Joomla CMS if we post a new tweet with media we receive error:

Deprecated: curl_setopt_array(): The usage of the @filename API for file uploading is deprecated. Please use the CURLFile class instead in libraries/joomla/http/transport/curl.php on line 179

This commit simply changes the code to use CURLFile.